refactor(model): summaries_set 下沉 EasyRecModel 以支持全部模型
将 build_summary_graph 从 RankModel 移至 EasyRecModel 基类，Match/召回等模型均可配置。
PCOC 新增 label_name 字段以支持多任务标签指定。
