+++ samples/model_config/deepfm_combo_on_avazu_ctr.config
@@ -371,6 +371,7 @@ model_config:{
   summaries_set {
     pcoc {
       # pred_name: "probs"   # prediction_dict key; "logits" applies sigmoid first
+      # label_name: "label"  # optional; for multi-task use e.g. probs_ctr + matching label
       # epsilon: 1e-7        # divisor stabilizer when observed ctr is near zero
     }
   }
